pyproject.toml
[tool.poetry]
name = "umk"
version = "0.0.3"
description = "Unimake is a set of development tools and frameworks for project maintaining. This tools makes it easy to organize development routines (such as building, testing, linting, running, etc)"
authors = ["Edward Sarkisyan <edw.sarkisyan@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-only"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

packages = [
    { include = "umk" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
rich = "^13.6.0"
gitpython = "^3.1.37"
asyncclick = "^8.1"
anyio = "^4.0.0"
python-dotenv = "^1.0.0"
beartype = "^0.16.3"
fs = "^2.4.16"
multimethod = "^1.10"
paramiko = "^3.3.1"
docker = "^7.0.0"
pydantic = "^2.5.3"

[tool.black]
line-length = 100


[tool.poetry.scripts]
umk = 'umk.tools.umk:entrypoint'
unimake = 'umk.tools.unimake:entrypoint'

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
